Move supabase config
