+++ lib/variables.ts
@@ -7,6 +7,9 @@ export const env = {
   private_key: process.env.GITHUB_PRIVATE_KEY,
   encryption_password: process.env.ENCRYPTION_PASSWORD,
   app_host: process.env.NEXT_PUBLIC_GISCUS_APP_HOST as `https://${string}`,
+  supabase_url: process.env.SUPABASE_URL,
+  supabase_table: process.env.SUPABASE_INSTALLATION_ACCESS_TOKENS_TABLE,
+  supabase_key: process.env.SUPABASE_KEY,
   revalidate_first_page: process.env.NEXT_PUBLIC_REVALIDATE_FIRST_PAGE !== 'false',
   origins: JSON.parse(process.env.ORIGINS || '[]') as string[],
   origins_regex: JSON.parse(process.env.ORIGINS_REGEX || '[]') as string[],
