+++ services/config.ts
@@ -15,13 +15,3 @@ export const GITHUB_REPO_INSTALLATION_URL = (repoWithOwner: string) =>
 
 export const GITHUB_ACCESS_TOKEN_URL = (id: number) =>
   `${GITHUB_INSTALLATIONS_URL}/${id}/access_tokens`;
-
-// Supabase
-
-export const SUPABASE_KEY = process.env.SUPABASE_KEY;
-
-const SUPABASE_URL = process.env.SUPABASE_URL;
-const SUPABASE_INSTALLATION_ACCESS_TOKENS_TABLE =
-  process.env.SUPABASE_INSTALLATION_ACCESS_TOKENS_TABLE || 'installation_access_tokens';
-
-export const SUPABASE_INSTALLATION_ACCESS_TOKENS_URL = `${SUPABASE_URL}/rest/v1/${SUPABASE_INSTALLATION_ACCESS_TOKENS_TABLE}`;
